+++ src/engine/src/low_level_renderer/vertex_buffer.cpp
@@ -6,8 +6,8 @@
 #include <glm/gtx/hash.hpp>
 
 #include "core/logger/assert.h"
-#include "private/buffer.h"
 #include "glm/gtx/hash.hpp"
+#include "private/buffer.h"
 
 namespace std {
 template <>
@@ -87,54 +87,76 @@ VertexBuffer VertexBuffer::create(
 	{  // populate vertices and indices arrays
 		std::unordered_map<Vertex, uint32_t> unique_vertices;
 		for (const auto& shape : shapes) {
-			int indexBegin = indices.size();
+			size_t indexBegin = indices.size();
 
-            ASSERT(shape.mesh.num_face_vertices[0] == 3, "We currently only support triangle meshes");
+			ASSERT(
+				shape.mesh.num_face_vertices[0] == 3,
+				"We currently only support triangle meshes"
+			);
 
-            for (size_t poly = 0, indexOffset = 0; poly < shape.mesh.num_face_vertices.size(); indexOffset += shape.mesh.num_face_vertices[poly++]) {
-                ASSERT(shape.mesh.num_face_vertices[poly] == 3, "Cannot support meshes with polygons of " << shape.mesh.num_face_vertices[poly] << " vertices, only triangle meshes permitted");
+			for (size_t poly = 0, indexOffset = 0;
+				 poly < shape.mesh.num_face_vertices.size();
+				 indexOffset += shape.mesh.num_face_vertices[poly++]) {
+				ASSERT(
+					shape.mesh.num_face_vertices[poly] == 3,
+					"Cannot support meshes with polygons of "
+						<< shape.mesh.num_face_vertices[poly]
+						<< " vertices, only triangle meshes permitted"
+				);
 
-                for (int faceIndex = 0; faceIndex < shape.mesh.num_face_vertices[poly]; faceIndex++) {
-                    const tinyobj::index_t index = shape.mesh.indices[indexOffset + faceIndex];
-                    ASSERT(index.normal_index >= 0, "Vertex does not specify a normal");
-                    ASSERT(index.texcoord_index >= 0, "Vertex does not specify a tex coordinate");
-                    const Vertex vertex{
-                        .position =
-                            glm::vec3{
-                                attrib.vertices[3 * index.vertex_index + 0],
-                                attrib.vertices[3 * index.vertex_index + 1],
-                                attrib.vertices[3 * index.vertex_index + 2]
-                            },
-                        .normal =
-                            glm::vec3{
-                                attrib.normals[3 * index.normal_index + 0],
-                                attrib.normals[3 * index.normal_index + 1],
-                                attrib.normals[3 * index.normal_index + 2]
-                            },
-                        .tangent = glm::vec3(0),
-                        .color = glm::vec3{1.0, 1.0, 1.0},
-                        // obj format coordinate system makes 0 the bottom of the
-                        // image, which is different from vulkan which considers it
-                        // the top
-                        .texCoord =
-                            glm::vec2{
-                                attrib.texcoords[2 * index.texcoord_index],
-                                1.0 - attrib.texcoords[2 * index.texcoord_index + 1]
-                            }
-                    };
+				for (size_t faceIndex = 0;
+					 faceIndex < shape.mesh.num_face_vertices[poly];
+					 faceIndex++) {
+					const tinyobj::index_t index =
+						shape.mesh.indices[indexOffset + faceIndex];
+					ASSERT(
+						index.normal_index >= 0,
+						"Vertex does not specify a normal"
+					);
+					ASSERT(
+						index.texcoord_index >= 0,
+						"Vertex does not specify a tex coordinate"
+					);
+					const Vertex vertex{
+						.position =
+							glm::vec3{
+								attrib.vertices[3 * index.vertex_index + 0],
+								attrib.vertices[3 * index.vertex_index + 1],
+								attrib.vertices[3 * index.vertex_index + 2]
+							},
+						.normal =
+							glm::vec3{
+								attrib.normals[3 * index.normal_index + 0],
+								attrib.normals[3 * index.normal_index + 1],
+								attrib.normals[3 * index.normal_index + 2]
+							},
+						.tangent = glm::vec3(0),
+						.color = glm::vec3{1.0, 1.0, 1.0},
+						// obj format coordinate system makes 0 the bottom of
+						// the image, which is different from vulkan which
+						// considers it the top
+						.texCoord =
+							glm::vec2{
+								attrib.texcoords[2 * index.texcoord_index],
+								1.0 -
+									attrib
+										.texcoords[2 * index.texcoord_index + 1]
+							}
+					};
 
-                    if (!unique_vertices.count(vertex)) {
-                        unique_vertices[vertex] =
-                            static_cast<uint32_t>(unique_vertices.size());
-                        vertices.push_back(vertex);
-                    }
-                    indices.push_back(unique_vertices[vertex]);
-                }
+					if (!unique_vertices.count(vertex)) {
+						unique_vertices[vertex] =
+							static_cast<uint32_t>(unique_vertices.size());
+						vertices.push_back(vertex);
+					}
+					indices.push_back(unique_vertices[vertex]);
+				}
 			}
 
-            constexpr int NUM_OF_VERTS_PER_TRIANGLE = 3;
+			constexpr int NUM_OF_VERTS_PER_TRIANGLE = 3;
 
-			for (int i = indexBegin; i < indices.size(); i += NUM_OF_VERTS_PER_TRIANGLE) {
+			for (size_t i = indexBegin; i < indices.size();
+				 i += NUM_OF_VERTS_PER_TRIANGLE) {
 				Vertex& v0 = vertices[indices[i]];
 				Vertex& v1 = vertices[indices[i + 1]];
 				Vertex& v2 = vertices[indices[i + 2]];
@@ -151,13 +173,13 @@ VertexBuffer VertexBuffer::create(
 				glm::vec3 tangent =
 					(deltaUV02.y * edge01 - deltaUV01.y * edge02) * normalizer;
 
-                for (int k = 0; k < NUM_OF_VERTS_PER_TRIANGLE; k++)
-                    vertices[indices[i + k]].tangent += tangent;
+				for (int k = 0; k < NUM_OF_VERTS_PER_TRIANGLE; k++)
+					vertices[indices[i + k]].tangent += tangent;
 			}
 		}
 
-        for (int i = 0; i < vertices.size(); i++)
-            vertices[i].tangent = glm::normalize(vertices[i].tangent);
+		for (size_t i = 0; i < vertices.size(); i++)
+			vertices[i].tangent = glm::normalize(vertices[i].tangent);
 	}
 
 	auto [vertexBuffer, deviceMemory] = Buffer::loadToBuffer(
